Fix containers not being accessible from 0.0.0.0
It is not possible to access the RabbitMQ port 5672 of the broker
container from other container (edge, server) trying to connect to it
through localhost. We need to use the wlp local IP address in ifconfig
(LOCAL_IP env variable).

This is now handled by CommUtils.parseRemoteIp, even if 0.0.0.0 is in
deployment_setup.json